+++ includes/MslsPlugin.php
@@ -217,7 +217,7 @@ class MslsPlugin {
 	 */
 	public function block_init() {
 		if ( ! $this->options->is_excluded() ) {
-            register_block_type( self::plugin_dir_path('dist/msls-widget-block' ) );
+            register_block_type( self::plugin_dir_path('js/msls-widget-block' ) );
             add_shortcode( 'sc_msls_widget', [ $this, 'block_render' ] );
 
 			return true;
@@ -251,15 +251,15 @@ class MslsPlugin {
 			return false;
 		}
 
-		$ver     = defined( 'MSLS_PLUGIN_VERSION' ) ? constant( 'MSLS_PLUGIN_VERSION' ) : false;
-		$postfix = defined( 'SCRIPT_DEBUG' ) && constant( 'SCRIPT_DEBUG' ) ? '' : '.min';
+		$ver    = defined( 'MSLS_PLUGIN_VERSION' ) ? constant( 'MSLS_PLUGIN_VERSION' ) : false;
+		$folder = defined( 'SCRIPT_DEBUG' ) && constant( 'SCRIPT_DEBUG' ) ? 'src' : 'js';
 
 		wp_enqueue_style( 'msls-styles', self::plugins_url( 'css/msls.css' ), [], $ver );
 		wp_enqueue_style( 'msls-flags', self::plugins_url( 'css-flags/css/flag-icon.min.css' ), [], $ver );
 
 		if ( $this->options->activate_autocomplete ) {
 			wp_enqueue_script( 'msls-autocomplete',
-				self::plugins_url( "js/msls{$postfix}.js" ),
+				self::plugins_url( "$folder/msls.js" ),
 				[ 'jquery-ui-autocomplete' ],
 				$ver );
