+++ js/msls-widget-block/index.asset.php
@@ -1,0 +1,1 @@
+<?php return array('dependencies' => array('react', 'wp-block-editor', 'wp-blocks', 'wp-components'), 'version' => '627848d7182ac425e62c');
