+++ package.json
@@ -1,8 +1,8 @@
 {
   "scripts": {
-    "uglify": "uglifyjs js/msls.js > js/msls.min.js",
+    "uglify": "uglifyjs src/msls.js > js/msls.js",
     "less": "lessc css/msls.less css/msls.css --clean-css=\"--s1 --advanced\"",
-    "build-msls-block": "wp-scripts build js/msls-widget-block/index.js --webpack-src-dir=js/msls-widget-block --output-path=dist/msls-widget-block",
+    "build-msls-block": "wp-scripts build src/msls-widget-block/index.js --webpack-src-dir=src/msls-widget-block --output-path=js/msls-widget-block",
     "build": "npm run uglify && npm run less && npm run build-msls-block"
   },
   "dependencies": {
