+++ src/msls-widget-block/block.json
@@ -1,0 +1,17 @@
+{
+  "$schema": "https://schemas.wp.org/trunk/block.json",
+  "apiVersion": 3,
+  "title": "Multisite Language Switcher",
+  "icon": "translation",
+  "category": "widgets",
+  "name": "lloc/msls-widget-block",
+  "version": "2.6.4",
+  "description": "This is the output form the_msls() that comes with the Multisite Language Switcher plugin as a Gutenberg block.",
+  "example": {},
+  "supports": {
+    "html": false
+  },
+  "textdomain": "multisite-language-switcher",
+  "editorScript": "file:./index.js",
+  "render": "file:./render.php"
+}
