+++ src/msls-widget-block/edit.js
