+++ src/msls-widget-block/index.js
@@ -1,0 +1,8 @@
+import { registerBlockType } from '@wordpress/blocks';
+
+import Edit from './edit';
+import metadata from './block.json';
+
+registerBlockType(metadata.name, {
+  edit: Edit
+} );
