+++ src/msls-widget-block/index.php
@@ -1,0 +1,3 @@
+<?php
+// Silence is golden.
+
