+++ src/msls-widget-block/render.php
@@ -1,0 +1,3 @@
+<?php
+
+the_msls();
